requirements.txt
django
PyOpenSSL==0.14
